Patch-10: ProcessLogger refactoring
- Moved ProcessLogger to ProcessingPlugin namespace
- Removed get methods from ProcessLogger interface
- Renamed table processes to link_pm_process_log
- Removed outdated ProcessViewController
