+++ src/Api/Log.php
@@ -1,0 +1,25 @@
+<?php
+/*
+ * This file is part of prooph/link.
+ * (c) 2014-2015 prooph software GmbH <contact@prooph.de>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ * 
+ * Date: 5/24/15 - 11:30 PM
+ */
+namespace Prooph\Link\ProcessManager\Api;
+use Prooph\Link\Application\Service\AbstractRestController;
+
+/**
+ * Resource Log
+ *
+ * This resource represents a process log of a performed workflow.
+ *
+ * @package Prooph\Link\ProcessManager\Api
+ * @author Alexander Miertsch <kontakt@codeliner.ws>
+ */
+final class Log extends AbstractRestController
+{
+
+}
