+++ src/ProcessingPlugin/ProcessLogger.php
@@ -9,7 +9,7 @@
  * Date: 20.01.15 - 22:51
  */
 
-namespace Prooph\Link\ProcessManager\Model;
+namespace Prooph\Link\ProcessManager\ProcessingPlugin;
 
 use Prooph\Processing\Processor\ProcessId;
 
@@ -69,28 +69,5 @@ interface ProcessLogger
      * @return void
      */
     public function logProcessFailed(ProcessId $processId, \DateTimeImmutable $finishedAt);
-
-    /**
-     * Orders process logs by started_at DESC
-     * Returns array of process log entry arrays.
-     * Each process log contains the information:
-     *
-     * - process_id => UUID string
-     * - status => running|succeed|failed
-     * - start_message => string|null
-     * - started_at => \DateTime::ISO8601 formatted
-     * - finished_at =>  \DateTime::ISO8601 formatted
-     *
-     * @param int $offset
-     * @param int $limit
-     * @return array
-     */
-    public function getLastLoggedProcesses($offset = 0, $limit = 10);
-
-    /**
-     * @param ProcessId $processId
-     * @return null|array process log, see {@method getLastLoggedProcesses} for structure
-     */
-    public function getLoggedProcess(ProcessId $processId);
 }
