+++ view/process-config/process-manager/riot-tag/process-flowchart-sidebar.phtml
@@ -1,7 +1,7 @@
 <div class="row">
     <div class="col-md-11 col-lg-offset-1">
         <ul class="nav nav-tabs">
-            <li class="active"><a href="#" class="process-tab" onclick="{ toggleActive }"><?php echo $this->translate('Process') ?></a></li>
+            <li class="active"><a href="#" class="process-tab" onclick="{ toggleActive }"><?php echo $this->translate('Workflow') ?></a></li>
             <li ><a href="#" class="elements-tab" onclick="{ toggleActive }"><?php echo $this->translate('Elements') ?></a></li>
         </ul>
         <div if="{ isProcessTabActive() }">
